Split Nuget for native and non-native apps: The single Microsoft.ApplicationInsights.ServiceFabric nuget package that we had before had a dependency on Microsoft.ServiceFabric nuget, in order to deal with ServiceContext from native apps.
The problem was that the ServiceFabric assemblies only work in x64 environment. This meant for lift and shift container scenarios even though the path to set ServiceContext was never getting hit, the app would crash if used on x86 environments (or for asp.net on IIS express from visual studio).

I am solving this by having two nugets, one for SF native apps specifi cally that has a dependency on Microsoft.ServiceFabric and a more generic one which can be used in all contexts since that doesn't depend on SF types.

In the long term the native nuget will be retired since we plan to move the functionality it provides inside the service fabric framework itself.